intermediate refactor
