+++ src/__init__.py
@@ -69,3 +69,10 @@ from .remove_remove_by_size_modifier import *
 
 from .sna_align_active_values_to_y import *
 
+from .sna_texture_creation import *
+from .apply_3dgs_transforms import *
+from .align_active_to_z_axis import *
+from .sna_viewport_render import *
+
+from .sna_shader_system import *
+from .export_mesh_object_as_3dgs_ply import *
