+++ src/auto_generate_crop_object.py
@@ -1,5 +1,4 @@
 import bpy
-from .important import *
 import bpy.utils.previews
 import webbrowser
 import os
