+++ src/important.py
@@ -18,6 +18,16 @@ from math import pi
 from mathutils import Matrix
 from typing import Optional
 
+# _icons = None
+#
+# def load_preview_icon(path):
+#     global _icons
+#     if not path in _icons:
+#         if os.path.exists(path):
+#             _icons.load(path, path, "IMAGE")
+#         else:
+#             return 0
+#     return _icons[path].icon_id
 
 def property_exists(prop_path, glob, loc):
     try:
